requirements.txt
asgiref==3.8.0
Django==5.0.3
django-tinymce==4.0.0
django-treebeard==4.7.1
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1